PDF_to_audio_translator/requirements.txt
Flask==2.2.5
pdf2image==1.16.3
pytesseract==0.3.10
PyPDF2==3.0.1
googletrans==4.0.0-rc1
gTTS==2.4.7
Pillow==9.5.0
Werkzeug==2.2.3
